% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{as_date_year_qtr}
\alias{as_date_year_qtr}
\title{Turn year quarter into date data}
\usage{
as_date_year_qtr(df, variable, format = "\%Y-Q\%q")
}
\arguments{
\item{df}{dataframe}

\item{variable}{field with data in format "2020-Q1"}

\item{format}{format to put date into}
}
\value{

}
\description{
Turn year quarter into date data
}
