% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{calc_unemployment_rate}
\alias{calc_unemployment_rate}
\title{Calculate unemployment metric}
\usage{
calc_unemployment_rate(data, var, ..., na.rm = TRUE)
}
\arguments{
\item{data}{tibble}

\item{var}{unemployment variable}

\item{...}{grouping variables}

\item{na.rm}{whether to remove null values from calculation, defaults to true}
}
\value{
tibble
}
\description{
Calculate unemployment metric
}
\examples{
\dontrun{
eoq::sample_cps \%>\% calc_unemployment_rate(employmentstatus)
eoq::sample_acs \%>\% calc_unemployment_rate(employmentstatus)
}
}
