% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{clean_education}
\alias{clean_education}
\title{Convert education data to title case}
\usage{
clean_education(data, var)
}
\arguments{
\item{data}{tibble}

\item{var}{education variable}
}
\value{
tibble
}
\description{
Convert education data to title case
}
\examples{
\dontrun{
eoq::sample_cps \%>\% clean_education(education)
eoq::sample_acs \%>\% clean_education(education)
}
}
