% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{clean_employment}
\alias{clean_employment}
\title{Clean unemployment variable}
\usage{
clean_employment(data, var)
}
\arguments{
\item{data}{a tibble}

\item{var}{employment variable name, typically `employmentstatus``}
}
\value{
a tibble
}
\description{
Clean unemployment variable
}
\examples{
\dontrun{
eoq::sample_cps \%>\% clean_employment(employmentstatus)
eoq::sample_acs \%>\% clean_employment(employmentstatus)
}
}
