% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{clean_race_ethnicity}
\alias{clean_race_ethnicity}
\title{Convert race / ethnicity to standard set}
\usage{
clean_race_ethnicity(data, var)
}
\arguments{
\item{data}{tibble}

\item{var}{race variable}
}
\value{
tibble
}
\description{
Convert race / ethnicity to standard set
}
\examples{
eoq::sample_cps \%>\% clean_race_ethnicity(racehispanic)
eoq::sample_acs \%>\% clean_race_ethnicity(racehispanic)
}
