% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{clean_sex}
\alias{clean_sex}
\title{Clean sex variable}
\usage{
clean_sex(data, var)
}
\arguments{
\item{data}{a tibble}

\item{var}{an is_male variable}
}
\value{
a tibble
}
\description{
Clean sex variable
}
