% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_qwi_metrics}
\alias{get_qwi_metrics}
\title{Get QWI metrics}
\usage{
get_qwi_metrics(
  name = "timeseries/qwi/se",
  vars = c(qwi_numeric_variables, "education"),
  region = "county:*",
  regionin = "state:06",
  time = "from 2018 to 2022",
  ...
)
}
\arguments{
\item{name}{}

\item{vars}{}

\item{region}{}

\item{regionin}{}

\item{time}{}

\item{...}{}
}
\value{

}
\description{
Get QWI metrics
}
