% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{make_age_buckets}
\alias{make_age_buckets}
\title{Make age buckets}
\usage{
make_age_buckets(data, var)
}
\arguments{
\item{data}{tibble}

\item{var}{age variable}
}
\value{
tibble
}
\description{
Make age buckets
}
\examples{
\dontrun{
eoq::sample_cps \%>\% make_age_buckets(age)
eoq::sample_acs \%>\% make_age_buckets(age)
}
}
