% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{make_percentiles}
\alias{make_percentiles}
\title{Make percentiles for chart}
\usage{
make_percentiles(data, var, ..., n = 100)
}
\arguments{
\item{data}{tibble}

\item{var}{variable to take the mean and median of}

\item{...}{grouping variables, optional}

\item{n}{number of ntiles, defaults to 100 for percentiles}
}
\value{
tibble
}
\description{
Make percentiles for chart
}
\examples{
\dontrun{
eoq::sample_acs \%>\% make_percentiles(totalwage)
eoq::sample_cps \%>\% make_percentiles(hourlywage, education)
}
}
