% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process cleaned ACS/CPS data for shiny app}
\usage{
process_data(data, input)
}
\arguments{
\item{data}{tibble}

\item{input}{input from shiny}
}
\value{
tibble
}
\description{
Process cleaned ACS/CPS data for shiny app
}
\examples{
\dontrun{
# only works inside a shiny app
eoq::sample_cps \%>\% process_data(input)
eoq::sample_acs \%>\% process_data(input)
}
}
