% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_individual_data}
\alias{process_individual_data}
\title{Process individual data}
\usage{
process_individual_data(data, input)
}
\arguments{
\item{data}{a tibble}

\item{input}{input from shiny}
}
\value{
tibble
}
\description{
Process individual data
}
