% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{safely_get_qwi_metrics}
\alias{safely_get_qwi_metrics}
\title{Safely get QWI metrics}
\usage{
safely_get_qwi_metrics(...)
}
\arguments{
\item{...}{fields to pass to get_qwi_metrics}
}
\value{

}
\description{
Safely get QWI metrics
}
