% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_acs}
\alias{sample_acs}
\title{Sample of ACS data}
\format{
A data frame with 32145 rows and 11 columns:
}
\source{
"~/Dropbox/Economic Opportunity Project/Data/Comparison to Peers/Outputs"
}
\usage{
sample_acs
}
\description{
A 1\% sample of the Clean_ACS.zip file in Dropbox
}
\keyword{datasets}
