% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cps}
\alias{sample_cps}
\title{Sample of CPS data}
\format{
A data frame with 42722 rows and 10 columns:
}
\source{
"~/Dropbox/Economic Opportunity Project/Data/Comparison to Peers/Outputs"
}
\usage{
sample_cps
}
\description{
A 3\% sample of the Clean_CPS.zip file in Dropbox
}
\keyword{datasets}
