% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{tidy_get_census}
\alias{tidy_get_census}
\title{get census tidy version}
\usage{
tidy_get_census(...)
}
\arguments{
\item{...}{fields to pass to censusapi::getCensus}
}
\value{

}
\description{
get census tidy version
}
