% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{as_date_year_qtr}
\alias{as_date_year_qtr}
\title{Title}
\usage{
as_date_year_qtr(df, variable, format = "\%Y-Q\%q", ...)
}
\arguments{
\item{df}{}

\item{variable}{}

\item{format}{}

\item{...}{}
}
\value{

}
\description{
Title
}
