% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_qwi_metrics}
\alias{get_qwi_metrics}
\title{Get QWI metrics}
\usage{
get_qwi_metrics(
  name = "timeseries/qwi/se",
  vars = c("Emp", "HirA", "HirAEndR", "SepBeg", "SepBegR", "EarnHirNS", "EarnS",
    "education", "sex"),
  region = "county:*",
  regionin = "state:06",
  time = "from 2018 to 2022",
  ...
)
}
\arguments{
\item{name}{name of census api}

\item{vars}{vars to select from api}

\item{region}{the level of geo (state, county)}

\item{regionin}{limiting the region to explore}

\item{time}{the time period for data}

\item{...}{addtional arguments passed to tidy_get_census}
}
\value{

}
\description{
Get QWI metrics
}
